// compiler/src/opcodes.d.ts
export declare enum OpCode {
    PushInt = 237,
    Pop = 33,
    Dup = 165,
    PushFloat = 109,
    PushString = 19,
    PushBool = 148,
    PushNull = 248,
    LoadLocal = 111,
    StoreLocal = 122,
    Add = 177,
    Sub = 16,
    Mul = 180,
    Div = 58,
    Eq = 32,
    Neq = 151,
    Lt = 181,
    Gt = 229,
    Lte = 91,
    Gte = 185,
    And = 121,
    Or = 100,
    Not = 62,
    BitAnd = 157,
    BitOr = 27,
    BitXor = 220,
    BitNot = 78,
    Shl = 69,
    Shr = 156,
    Jump = 87,
    JumpIf = 59,
    JumpIfNot = 158,
    NewObject = 129,
    NewList = 143,
    ListPush = 135,
    GetMember = 42,
    SetMember = 178,
    Length = 95,
    Hash256 = 74,
    EncryptAES = 245,
    JSONStringify = 250,
    Call = 255,
    Return = 173,
    CallNative = 234,
    Halt = 228,
    Concat = 107,
    CompareAndAdd = 241,
    SwapAndMul = 141,
    JumpAndMul = 202,
    Swap = 254,
    Rotate = 236,
    Drop2 = 218
}
//# sourceMappingURL=opcodes.d.ts.map